moved error tracker
